# api.yaml
openapi: 3.0.0
servers:
  - url: http://localhost:8080
info:
  title: CIRCABC Share
  description: This is a API definition for the CIRCABC Share service.
  license:
    name: 'EUPL V1.2'
    url: 'https://joinup.ec.europa.eu/sites/default/files/inline-files/EUPL%20v1_2%20EN(1).txt'
  version: '0.1'
...
paths:
...
  '/user/{userID}/userInfo':
    parameters:
      - name: userID
        in: path
        description: The id of the user
        required: true
        schema:
          type: string 
    put:
      description: Used by the administrators in order to update a specific INTERNAL user total space or admin status
      operationId: putUserUserInfo
      parameters:
        - name: freezeAll
          in: query
          description: Freeze or unfreeze user's files
          required: false
          schema:
            type: boolean
      tags:
      - Users
      security:
        - OpenID:
          - email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInfo'
      responses:
        200:
          description: SUCCESS Returns the UserInfo of the selected user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        400:
          description: BAD REQUEST the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        401:
          description: UNAUTHORIZED the Error message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        403:
          description: FORBIDDEN the Error message will be NotAuthorized, will be sent before 404
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        404:
          description: NOT FOUND the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        500:
          description: INTERNAL SERVER ERROR the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    get:
      description: Used by the users in order to fetch their personal information
      operationId: getUserUserInfo
      tags:
      - Users
      security:
        - OpenID:
          - email
      responses:
        200:
          description: SUCCESS Returns the UserInfo of the selected user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        401:
          description: UNAUTHORIZED the Error message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        403:
          description: FORBIDDEN the Error message will be NotAuthorized, will be sent before 404
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        404:
          description: NOT FOUND the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        500:
          description: INTERNAL SERVER ERROR the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  '/user/{userID}/files/fileInfoUploader':
    parameters:
      - name: userID
        in: path
        description: The id of the user
        required: true
        schema:
          type: string 
      - name: pageSize
        in: query
        description: Number of files returned
        required: true
        schema:
          type: integer
      - name: pageNumber
        in: query
        description: Page number
        required: true
        schema:
          type: integer
    get:
      description: Used by the INTERNAL users in order to search their own files' FileInfoUploader
      operationId: getFilesFileInfoUploader
      tags:
        - Users
      security:
        - OpenID:
          - email
      responses:
        200:
          description: SUCCESS Returns the files shared with the authenticated user by pageSize and pageNumber.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileInfoUploader'
        400:
          description: BAD REQUEST the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        401:
          description: UNAUTHORIZED the Error message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        403:
          description: FORBIDDEN the Error message will be NotAuthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        404:
          description: NOT FOUND the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        500:
          description: INTERNAL SERVER ERROR the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  '/user/{userID}/files/fileInfoRecipient':
    parameters:
      - name: userID
        in: path
        description: The id of the user
        required: true
        schema:
          type: string 
      - name: pageSize
        in: query
        description: Number of files returned
        required: true
        schema:
          type: integer
      - name: pageNumber
        in: query
        description: Page number
        required: true
        schema:
          type: integer
    get:
      description: Used by the INTERNAL users in order to search the files they have recieved
      operationId: getFilesFileInfoRecipient
      tags:
        - Users
      security:
        - OpenID:
          - email
      responses:
        200:
          description: SUCCESS Returns the files shared with the authenticated user by pageSize and pageNumber
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileInfoRecipient'
        400:
          description: BAD REQUEST the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        401:
          description: UNAUTHORIZED the Error message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        403:
          description: FORBIDDEN the Error message will be NotAuthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        404:
          description: NOT FOUND the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        500:
          description: INTERNAL SERVER ERROR the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
components:
  securitySchemes:
    OpenID:   # <--- Arbitrary name for the security scheme. Used to refer to it from elsewhere.
      type: openIdConnect
      openIdConnectUrl: https://localhost:8080/realms/
...
  schemas:
    Status:
      type: object
      properties:
        code:
          type: integer
          description: Status code
        message:
          type: string
          description: Additional message
      required:
        - code
    FileInfoUploader:
      allOf:
        - $ref: '#/components/schemas/FileBasics'
        - type: object
          properties:
            fileId:
              description: file id
              type: string
            sharedWith:
              description: User IDs this file is shared with
              type: array
              items:
                $ref: '#/components/schemas/Recipient'
              minItems: 1
              maxItems: 10
            fileLogs:
              description: File logs
              type: array
              items:
                $ref: '#/components/schemas/FileLog' 
            status:
              type: string
              enum: 
                - "AVAILABLE"
                - "ALLOCATED"
                - "DELETED"
                - "FROZEN"
                - "UPLOADING"
          required:
            - sharedWith
            - fileLogs
            - fileId
    FileInfoRecipient:
      allOf:
        - $ref: '#/components/schemas/FileBasics'
        - type: object
          properties:
            uploaderName:
              description: name of the uploader
              type: string
            fileId:
              description: download file id
              type: string

          required:
            - uploaderName
            - fileId
    FileBasics:
      type: object
      properties:
        expirationDate:
          description: Expiration date of file
          type: string
          format: date
        hasPassword:
          description: File is password-protected
          type: boolean
        name:
          description: Filename
          type: string
        size:
          description: Size of file (Bytes)
          type: number
          minimum: 0
      required:
      - expirationDate
      - hasPassword
      - name
      - size
    FileLog:
      type: object
      properties:
        fileId:
          description: ID of the file
          type: string
        downloadDate:
          description: Download date of file
          type: string
        recipient:
          description: recipient of the file
          type: string
        downloadLink:
          description: download link
          type: string
      required:
      - fileId
      - downloadDate
      - recipient
      - downloadLink
...
    Recipient:
      properties:
        email:
          description: Email of the recipient
          type: string
        message: 
          description: Optional message to send
          type: string
        downloadLink:
          description: Download link to a specific file
          type: string
        shortUrl:
          description: Short link to a specific file
          type: string
        downloadNotification:
          description: Email notification sent after download if set to true
          type: boolean
    UserInfo:
      allOf:
        - $ref: '#/components/schemas/UserSpace'
        - type: object
          properties:
            id: 
              description: User ID 
              type: string
            loginUsername:
              description: Abbreviated user name used for login
              type: string
            givenName:
              description: Full name of the user
              type: string
            email:
              description: Email address 
              type: string
            role:
              description: role
              type: string
              enum:
               - ADMIN
               - INTERNAL
               - EXTERNAL
               - TRUSTED_EXTERNAL
               - API_KEY
            isAdmin:
              description: True if the user is admin
              type: boolean
            status:
              description: status = REGULAR (keep all user's rights) - Users who have full access to the site; SUSPENDED (Prevent users from uploading) - Users who can log in and download, but cannot upload; BANNED (prevent user from login) - Users who cannot log in at all.
              type: string
              enum:
                - REGULAR
                - SUSPENDED
                - BANNED
          required:
          - id
          - loginUsername
          - email
          - isAdmin
    UserSpace:
      type: object
      properties:
        totalSpace:
          description: Total space the user has (Bytes)
          type: number
          minimum: 0
        usedSpace:
          description: Space the user already used (Bytes)
          type: number
          minimum: 0
        filesCount:
          description: Number of files
          type: number
          minimum: 0
      required:
      - totalSpace
      - usedSpace
...